IntroductionToComputationalBiology/requirements.txt
biopython==1.78
goatools==1.1.5
networkx==2.5.1
numpy==1.19.5
obonet==0.3.0
scipy==1.6.3
statsmodels==0.12.2
